+++ apps/faucet/src/components/ProcessingModal.tsx
@@ -1,31 +1,44 @@
+import getExplorerURI from '@webb-tools/api-provider-environment/transaction/utils/getExplorerURI';
 import { chainsConfig } from '@webb-tools/dapp-config/chains/chain-config';
-import { MetaMaskIcon } from '@webb-tools/icons';
+import isValidUrl from '@webb-tools/dapp-types/utils/isValidUrl';
+import { MetaMaskIcon, WalletLineIcon } from '@webb-tools/icons';
 import {
   calculateTypedChainId,
   ChainType,
 } from '@webb-tools/sdk-core/typed-chain-id';
 import {
   Button,
+  KeyValueWithButton,
   Modal,
   ModalContent,
   ModalFooter,
   ModalHeader,
+  populateDocsUrl,
   Typography,
 } from '@webb-tools/webb-ui-components';
+import {
+  BRIDGE_URL,
+  GITHUB_BUG_REPORT_URL,
+  WEBB_DOC_ROUTES_RECORD,
+} from '@webb-tools/webb-ui-components/constants';
 import cx from 'classnames';
 import Lottie from 'lottie-react';
 import { useObservableState } from 'observable-hooks';
-import { useCallback, useMemo } from 'react';
+import { type ComponentProps, useCallback, useMemo } from 'react';
 
 import FaucetError from '../errors/FaucetError';
-import FaucetErrorCode from '../errors/FaucetErrorCode';
-import ErrorPayload from '../errors/FaucetErrorPayload';
 import failedAnimation from '../lottie/failed.json';
 import processingAnimation from '../lottie/processing.json';
 import sucessAnimation from '../lottie/success.json';
 import { useFaucetContext } from '../provider';
+import type { MintTokenBody, MintTokenResult } from '../types';
 import addTokenToMetamask from '../utils/addTokenToMetamask';
 import parseErrorFromResult from '../utils/parseErrorFromResult';
+
+const sharedExternalLinkProps = {
+  rel: 'noopener noreferrer',
+  target: '_blank',
+} satisfies ComponentProps<'a'>;
 
 const ProcessingModal = () => {
   const {
@@ -57,58 +70,6 @@ const ProcessingModal = () => {
         : processingAnimation,
     [isSuccess, isFailed]
   );
-
-  const mintTxLink = useMemo(() => {
-    if (!mintTokenRes) return '';
-
-    if (FaucetError.isFaucetError(mintTokenRes)) return '';
-
-    const {
-      tx_result: { Evm: txReceipt, Substrate: txHash },
-      typed_chain_id: { Evm: evmChainId, Substrate: substrateChainId },
-    } = mintTokenRes;
-
-    let typedChainId: number;
-
-    if (substrateChainId) {
-      typedChainId = calculateTypedChainId(
-        ChainType.Substrate,
-        substrateChainId
-      );
-    } else if (evmChainId) {
-      typedChainId = calculateTypedChainId(ChainType.EVM, evmChainId);
-    } else {
-      alert('No chain id found in the mint token result');
-      return '';
-    }
-
-    const chain = chainsConfig[typedChainId];
-
-    if (!chain) {
-      alert(`Typed chain id ${typedChainId} is not in the chains config`);
-      return '';
-    }
-
-    if (!chain.blockExplorers) {
-      alert(`Chain ${chain.name} does not have a block explorer url`);
-      return '';
-    }
-
-    if (txReceipt) {
-      const explorerUrl = chain.blockExplorers.default.url;
-
-      return new URL(
-        `/tx/${txReceipt.transactionHash}`,
-        explorerUrl
-      ).toString();
-    } else if (txHash) {
-      alert(`Substrate tx hash ${txHash} is not supported yet`);
-      return '';
-    } else {
-      alert('No tx hash found in the mint token result');
-      return '';
-    }
-  }, [mintTokenRes]);
 
   const errorMessage = useMemo(
     () => parseErrorFromResult(mintTokenRes),
@@ -169,42 +130,59 @@ const ProcessingModal = () => {
 
           <Typography fw="semibold" ta="center" variant="body1">
             {isSuccess
-              ? 'This transfer has been made to your wallet address.'
+              ? getSuccesMessage()
               : isFailed
               ? errorMessage
               : 'Your request is in progress. It may take up to a few seconds to complete the request.'}
           </Typography>
 
-          <Button
-            href={mintTxLink}
-            target="_blank"
-            rel="noopener noreferrer"
-            className={cx({ hidden: !isSuccess })}
-            variant="link"
-          >
-            View on Explorer
-          </Button>
+          <MintTxLinkOrHash mintTokenResult={mintTokenRes} />
         </div>
 
         {/* Hide the footer while transaction is in-progress */}
         <ModalFooter className={cx({ hidden: !isSuccess && !isFailed })}>
           {isSuccess ? (
-            <Button
-              rightIcon={<MetaMaskIcon size="lg" />}
-              onClick={handleAddTokenToMetamask}
-              isFullWidth
-              variant="secondary"
-            >
-              Add token to Metamask
-            </Button>
+            <>
+              <Button
+                {...sharedExternalLinkProps}
+                href={BRIDGE_URL}
+                isFullWidth
+              >
+                Hubble Bridge
+              </Button>
+              <Button
+                rightIcon={
+                  inputValues.recepientAddressType === 'ethereum' ? (
+                    <MetaMaskIcon size="lg" />
+                  ) : (
+                    <WalletLineIcon size="lg" />
+                  )
+                }
+                onClick={handleAddTokenToMetamask}
+                isFullWidth
+                variant="secondary"
+                isDisabled={inputValues.recepientAddressType === 'substrate'}
+              >
+                Add token to{' '}
+                {inputValues.recepientAddressType === 'ethereum'
+                  ? 'MetaMask'
+                  : 'wallet'}
+              </Button>
+            </>
           ) : (
-            <Button
-              onClick={() => handleOpenChange(false)}
-              isFullWidth
-              variant="secondary"
-            >
-              Close
-            </Button>
+            <>
+              <Button onClick={() => handleOpenChange(false)} isFullWidth>
+                Close
+              </Button>
+              <Button
+                {...sharedExternalLinkProps}
+                isFullWidth
+                variant="secondary"
+                href={GITHUB_BUG_REPORT_URL}
+              >
+                Report Bug
+              </Button>
+            </>
           )}
         </ModalFooter>
       </ModalContent>
@@ -213,3 +191,115 @@ const ProcessingModal = () => {
 };
 
 export default ProcessingModal;
+
+const MintTxLinkOrHash = (props: {
+  mintTokenResult: MintTokenResult | null;
+}) => {
+  const { mintTokenResult } = props;
+
+  const txLinkOrTxHash = useMemo(() => {
+    if (!mintTokenResult) return '';
+
+    if (FaucetError.isFaucetError(mintTokenResult)) return '';
+
+    const { hash, typedChainId, isSubstrate } =
+      parseMintResult(mintTokenResult);
+
+    const chain = chainsConfig[typedChainId];
+
+    if (!chain) {
+      console.warn(
+        `Typed chain id ${typedChainId} is not in the chains config`
+      );
+      return hash;
+    }
+
+    if (!chain.blockExplorers || !hash) {
+      return hash;
+    }
+
+    try {
+      return getExplorerURI(
+        chain.blockExplorers.default.url,
+        hash,
+        'tx',
+        isSubstrate ? 'polkadot' : 'web3'
+      ).toString();
+    } catch (error) {
+      console.error(error);
+      return hash;
+    }
+  }, [mintTokenResult]);
+
+  if (!txLinkOrTxHash) return null;
+
+  if (isValidUrl(txLinkOrTxHash)) {
+    return (
+      <Button
+        {...sharedExternalLinkProps}
+        href={txLinkOrTxHash}
+        isDisabled={!txLinkOrTxHash}
+        variant="link"
+      >
+        View on Explorer
+      </Button>
+    );
+  }
+
+  return (
+    <KeyValueWithButton
+      label="Transaction Hash"
+      size="sm"
+      className="mx-auto"
+      keyValue={txLinkOrTxHash}
+    />
+  );
+};
+
+const usageUrl = populateDocsUrl(
+  WEBB_DOC_ROUTES_RECORD['projects']['hubble-bridge']['usage-guide'].route
+);
+
+const getSuccesMessage = () => {
+  return (
+    <>
+      This transfer has been made to your wallet address. Experience private
+      bridging on Hubble or explore token usages on the{' '}
+      <a
+        className="!text-inherit hover:underline"
+        {...sharedExternalLinkProps}
+        href={usageUrl}
+      >
+        doc site
+      </a>
+      !
+    </>
+  );
+};
+
+const parseMintResult = (
+  result: MintTokenBody
+): {
+  isSubstrate: boolean;
+  hash: string;
+  typedChainId: number;
+} => {
+  const { typed_chain_id, tx_result } = result;
+
+  if ('Substrate' in tx_result) {
+    return {
+      hash: tx_result.Substrate.block_hash,
+      isSubstrate: true,
+      typedChainId: calculateTypedChainId(
+        ChainType.Substrate,
+        typed_chain_id.id
+      ),
+    };
+  }
+
+  return {
+    hash: tx_result.Evm.transactionHash,
+    isSubstrate: false,
+    typedChainId: calculateTypedChainId(ChainType.EVM, typed_chain_id.id),
+  };
+};