+++ apps/hubble-stats/containers/PoolInfoCardContainer/PoolInfoCardContainer.tsx
@@ -9,9 +9,16 @@ import {
   PoolTypeChip,
 } from '../../components';
 import { VANCHORS_MAP } from '../../constants';
-import { getPoolInfoCardTvlData, getPoolInfoCardDepositData } from '../../data';
+import {
+  getPoolInfoCardTvlData as getPoolTvl,
+  getPoolInfoCardDepositData as getPoolDeposit,
+} from '../../data';
+import { cache } from 'react';
 
-export default async function PoolInfoCardContainer({
+const getPoolInfoCardTvlData = cache(getPoolTvl);
+const getPoolInfoCardDepositData = cache(getPoolDeposit);
+
+export default function PoolInfoCardContainer({
   poolAddress,
   epochStart,
   epochNow,