+++ libs/dapp-types/src/utils/index.ts
@@ -1,3 +1,4 @@
+export { default as isSubstrateAddress } from './isSubstrateAddress';
 export { default as isValidAddress } from './isValidAddress';
 export { default as isValidPublicKey } from './isValidPublicKey';
 export { default as isValidUrl } from './isValidUrl';